refactor: 폴더 위치 변경 및 불필요한 파일 삭제
