+++ src/hooks/useResetPW.ts
@@ -1,5 +1,5 @@
 import { useMutation } from "@tanstack/react-query";
-import { put } from "@/api/http";
+import { put } from "@/util/http";
 
 interface ResetPasswordData {
   userid: string;
