+++ src/hooks/useTabs.ts
@@ -1,7 +1,7 @@
 import { Tab } from "@/models/tab.model";
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 import { queryKeys } from "@/constants/queryKeys";
-import { get, put } from "@/api/http";
+import { get, put } from "@/util/http";
 import { useState } from "react";
 import { UpdateContentsRequest } from "@/app/api/contents/route";
 import { useTranslations } from "next-intl";
