+++ src/hooks/useUser.ts
@@ -1,7 +1,7 @@
 import { User } from "@/models/user.model";
 import { useQuery } from "@tanstack/react-query";
 import { queryKeys } from "@/constants/queryKeys";
-import { get } from "@/api/http";
+import { get } from "@/util/http";
 
 export const useUser = (userid: string) => {
   const {
